pyproject.toml
[tool.poetry]
name = "neoplasm-users"
version = "0.1.0"
description = "Neoplasm Users service"
license = "MIT"
authors = ["Pavel <pavel@neoplasm.xyz>"]
readme = "README.md"
repository = "https://github.com/nplsm/neoplasm-users"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = {version = "^0.12.1", extras = ["standard"]}
ariadne = "^0.12.0"
pydantic = {version = "^1.6.1", extras = ["email"]}
motor = "^2.3.0"
python-jose = {version = "^3.2.0", extras = ["cryptography"]}
passlib = {version = "^1.7.4", extras = ["argon2"]}

[tool.poetry.dev-dependencies]
isort = "^5.6.4"
rope = "^0.18.0"
black = "^20.8b1"
flake8 = "^3.8.4"
flake8-bugbear = "^20.1.4"
flake8-pytest = "^1.3"
flake8-pytest-style = "^1.3.0"
mypy = "^0.790"
pytest = "^6.1.1"
pytest-cov = "^2.10.1"
pre-commit = "^2.7.1"
requests = "^2.24.0"

[build-system]
requires = [
    "poetry-core>=1.0.0",
    "poetry>=1.1.3"
]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = [
    "tests",
]

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
